Crate aster_block

The block devices of Asterinas. an abstraction of block devices, as well as the registration and lookup of block devices.

Block devices use a queue-based model for asynchronous I/O operations. It is necessary for a block device to maintain a queue to handle I/O requests. The users (e.g., fs) submit I/O requests to this queue and wait for their completion. Drivers implementing block devices can create their own queues as needed, with the possibility to reorder and merge requests within the queue.

This crate also offers the Bio related data structures and APIs to accomplish safe and convenient block I/O operations, for example:

// Creates a bio request.
let bio = Bio::new(BioType::Write, sid, segments, None);
// Submits to the block device.
let bio_waiter = bio.submit(block_device)?;
// Waits for the the completion.
let Some(status) = bio_waiter.wait() else {
    return Err(IoError);
};
assert!(status == BioStatus::Complete);

Modules

bio

id

request_queue

Structs

BlockDeviceMeta

Metadata for a block device.

MajorIdOwner

An owned major ID.

PartitionNode

Enums

Error

The error type which is returned from the APIs of this crate.

PartitionInfo

Represents a partition entry.

Constants

BLOCK_SIZE

SECTOR_SIZE

Statics

EXTENDED_DEVICE_ID_ALLOCATOR

Traits

BlockDevice

Functions

acquire_major

Acquires a major ID.

allocate_major

Allocates a major ID.

collect_all

Collects all block devices.

lookup

Looks up a block device of a given device ID.

register

Registers a new block device.

unregister

Unregisters an existing block device, returning the device if found.