aster_block

Module bio

Structs

AlignedUsize

An aligned unsigned integer number.

Bio

The unit for block I/O.

BioSegment

BioSegment is the basic memory unit of a block I/O request.

BioWaiter

A waiter for Bio submissions.

SubmittedBio

A submitted Bio object.

Enums

BioDirection

The direction of a bio request.

BioEnqueueError

The error type returned when enqueueing the Bio.

BioStatus

The status of Bio.

BioType

The type of Bio.

Functions

bio_segment_pool_init

Initializes the bio segment pool.

is_sector_aligned

Checks if the given offset is aligned to sector.