aster_block::id

Struct BlockId

pub struct BlockId<const N: u16>(/* private fields */);

An index of a block.

The BlockId<const N: u16> is a generic type that is parameterized by a constant N, which represents the size of each block in bytes. The BlockId<_> provides a type-safe way of handling block indices. An Instance of BlockId<_> is guaranteed to represent valid block index, derived from byte offset and the specified block size N.

Examples

const BLOCK_SIZE: u16 = 512;

let bytes_offset = 2048;
let block_id = BlockId<BLOCK_SIZE>::from_offset(bytes_offset);
assert!(block_id == (bytes_offset / BLOCK_SIZE));

Limitation

Currently, the block size is expressed in u16. We choose u16 because it is reasonably large to represent the common block size used in practice.

Implementations

impl<const N: u16> BlockId<N>

pub const fn new(raw_id: u64) -> Self

Constructs an id from a raw id.

pub const fn from_offset(offset: usize) -> Self

Constructs an id from a byte offset.

pub fn to_offset(self) -> usize

Converts to a byte offset.

pub fn to_raw(self) -> u64

Converts to raw id.